pyproject.toml
[project]
name = "p2-requetez-des-services-ia"
version = "0.1.0"
description = ""
authors = [
    {name = "Jody Gs",email = "jodygonzales.pro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
